src/bot/requirements-dev.txt
autopep8==1.2.4

pytest==3.0.2
pytest-aiohttp==0.1.3
pytest-watch==4.1.0
